Refactor tests to tests/ folder and enable recursive pylint
Moved `run_real_import_test.py`, `test_import.py`, and `sample.mbox` to `tests/` directory.
Updated `tests/test_import.py` and `tests/run_real_import_test.py` to correctly reference the main script and test data using relative paths.
Updated `.github/workflows/ci.yml` to run pylint recursively on all python files and discover tests in `tests/`.
Updated `build.py` to discover tests in `tests/` and fixed lint errors (indentation, docstrings).
Fixed lint errors in `tests/test_import.py` (indentation, docstrings).
