+++ build.py
@@ -1,34 +1,38 @@
+"""Build and test script for import-mailbox-to-gmail."""
 import unittest
 import subprocess
 import sys
 
 def run_tests():
-    """Runs the unit tests."""
-    loader = unittest.TestLoader()
-    suite = loader.discover('.')
-    runner = unittest.TextTestRunner()
-    result = runner.run(suite)
-    return len(result.failures) == 0 and len(result.errors) == 0
+  """Runs the unit tests."""
+  loader = unittest.TestLoader()
+  suite = loader.discover('tests', top_level_dir='.')
+  runner = unittest.TextTestRunner()
+  result = runner.run(suite)
+  return len(result.failures) == 0 and len(result.errors) == 0
 
 def create_executable():
-    """Creates the executable using PyInstaller."""
+  """Creates the executable using PyInstaller."""
+  try:
+    import PyInstaller  # pylint: disable=unused-import,import-outside-toplevel
+  except ImportError:
+    print("PyInstaller not found, installing...")
     try:
-        import PyInstaller
-    except ImportError:
-        print("PyInstaller not found, installing...")
-        try:
-            subprocess.check_call([sys.executable, '-m', 'pip', 'install', 'pyinstaller'])
-        except subprocess.CalledProcessError as e:
-            print(f"Failed to install PyInstaller: {e}")
-            sys.exit(1)
+      subprocess.check_call([sys.executable, '-m', 'pip', 'install', 'pyinstaller'])
+    except subprocess.CalledProcessError as e:
+      print(f"Failed to install PyInstaller: {e}")
+      sys.exit(1)
 
-    print("Running PyInstaller...")
-    subprocess.run([sys.executable, '-m', 'PyInstaller', '--onefile', 'import-mailbox-to-gmail.py'])
+  print("Running PyInstaller...")
+  subprocess.run(
+      [sys.executable, '-m', 'PyInstaller', '--onefile', 'import-mailbox-to-gmail.py'],
+      check=True
+  )
 
 if __name__ == '__main__':
-    if run_tests():
-        print("Tests passed, creating executable...")
-        create_executable()
-    else:
-        print("Tests failed, not creating executable.")
-        sys.exit(1)
+  if run_tests():
+    print("Tests passed, creating executable...")
+    create_executable()
+  else:
+    print("Tests failed, not creating executable.")
+    sys.exit(1)
